+++ src/leksem_creator.h
@@ -34,7 +34,7 @@
 #define SUPREME_PRIORITY 3
 
 typedef struct lexeme {
-  int value;
+  double value;
   double double_num;
   int value_type;
   int priority;
@@ -47,9 +47,9 @@ int handle_variable(char **input_str, struct lexeme *output_lexeme);
 int one_symbol_operator(char **input_str, struct lexeme *output_lexeme);
 int multi_symbol_operator(char **input_str, struct lexeme *output_lexeme);
 void add_number(char **input_str, struct lexeme *output_lexeme);
-void read_number(char **input_str, int *int_number, int *int_float);
-void write_number(struct lexeme *output_lexeme, int int_number, int int_float);
-int converter_str_to_int(char **input_str);
+void read_number(char **input_str, double *int_number, double *int_float);
+void write_number(struct lexeme *output_lexeme, double int_number, double int_float);
+double converter_str_to_int(char **input_str);
 void lexemes_initiate(struct lexeme *output_lexeme);
 
 #endif  // LEXEME_CREATOR_H_
